Convert Upgrades Context to Zustand
