+++ frontend/app/components/BlockView.tsx
@@ -2,7 +2,7 @@ import { useCallback, useEffect, useState } from "react";
 import { View, Text, Image, LayoutChangeEvent } from "react-native";
 import Animated, { BounceIn, Easing, FadeOut } from "react-native-reanimated";
 import messagesJson from "../configs/messages.json";
-import { useUpgrades } from "../context/Upgrades";
+import { useUpgrades } from "../stores/useUpgradesStore";
 import { useImages } from "../hooks/useImages";
 import { useTutorialLayout } from "../hooks/useTutorialLayout";
 import { TargetId } from "../stores/useTutorialStore";
