+++ frontend/app/components/WorkingBlockView.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from "react";
 import { StyleProp, Text, View, ViewStyle } from "react-native";
 import { useGameStore } from "@/app/stores/useGameStore";
-import { useUpgrades } from "../context/Upgrades";
+import { useUpgrades } from "../stores/useUpgradesStore";
 import { useImages } from "../hooks/useImages";
 import { BlockView } from "./BlockView";
 import { Miner } from "./Miner";
