+++ frontend/app/components/store/PrestigeUnlock.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 import { View } from "react-native";
-import { useUpgrades } from "../../context/Upgrades";
+import { useUpgrades } from "../../stores/useUpgradesStore";
 import { UnlockView } from "./UnlockView";
 
 export const PrestigeUnlock: React.FC = () => {
