+++ frontend/app/pages/LeaderboardPage.tsx
@@ -15,7 +15,7 @@ import { LinearGradient } from "expo-linear-gradient";
 import { PFPView } from "../components/PFPView";
 import { useStarknetConnector } from "../context/StarknetConnector";
 import { useBalance } from "../stores/useBalanceStore";
-import { useUpgrades } from "../context/Upgrades";
+import { useUpgrades } from "../stores/useUpgradesStore";
 import { useFocEngine } from "../context/FocEngineConnector";
 import { useImages } from "../hooks/useImages";
 import { usePowContractConnector } from "../context/PowContractConnector";
